Move from mechanize to grab.
There is an open issue since May to move mechanize to Python3,
but it seems to take longer until it will reach this milestone
(https://github.com/python-mechanize/mechanize/issues).

So instead of waiting for mechanize to gain Python3 support, this
moves the whole functionality to grab
(https://github.com/lorien/grab/issues), which provides similar
functionality as mechanize but supports Python2 and Python3.